refactor: move list user games
