+++ server/internal/transport/http/game.go
@@ -55,8 +55,8 @@ func (h *handler) ListGames(w http.ResponseWriter, r *http.Request, params api.L
 	writeResponseJSON(w, http.StatusOK, responseBody)
 }
 
-// ListRecommendedGames handles the http request to list recommended games.
-func (h *handler) ListRecommendedGames(w http.ResponseWriter, r *http.Request, params api.ListRecommendedGamesParams) {
+// ListGamesRecommended handles the http request to list recommended games.
+func (h *handler) ListGamesRecommended(w http.ResponseWriter, r *http.Request, params api.ListGamesRecommendedParams) {
 	ctx := r.Context()
 
 	domainGamesRecommendedFilter := listGamesRecommendedParamsToDomain(params)
@@ -285,8 +285,8 @@ func listGamesParamsToDomain(params api.ListGamesParams) domain.GamesPaginatedFi
 }
 
 // listGamesRecommendedParamsToDomain returns a domain games recommended paginated filter based on the standardized list
-// recommended games parameters.
-func listGamesRecommendedParamsToDomain(params api.ListRecommendedGamesParams) domain.GamesRecommendedPaginatedFilter {
+// games recommended parameters.
+func listGamesRecommendedParamsToDomain(params api.ListGamesRecommendedParams) domain.GamesRecommendedPaginatedFilter {
 	return domain.GamesRecommendedPaginatedFilter{
 		PaginatedRequestBase: paginatedRequestBaseToDomain(
 			params.Limit,
